// frontend/src/types/overview.ts
export interface Overview {

    totalMachines: number;

    runningMachines: number;

    warningMachines: number;

    criticalMachines: number;

    offlineMachines: number;

    activeAlerts: number;

    averageTemperature: number;

    averageHealthScore: number;

    totalProduction: number;

}